refactor: Cursor ResponseDto 공용으로 쓸 수 있도록 위치 이동 (global > common > pagination)
